changed routes to pages. Consolidating components into features. Auth feature component done. Continuing.
