+++ src/features/auth/Auth.tsx
@@ -1,0 +1,11 @@
+import LoginContainer from "./containers/AuthContainer.Login";
+import LoginForm from "./components/LoginForm";
+
+export default function Auth() {
+
+  return (
+    <LoginContainer>
+      <LoginForm />
+    </LoginContainer>
+  )
+}
