+++ src/main.tsx
@@ -9,12 +9,12 @@ import {
   createRoutesFromElements,
 } from "react-router-dom";
 
-import Home from './routes/home.tsx';
+import Home from './pages/home.tsx';
 // import PracticeContainer from "containers/practice/PracticeContainer";
-import Practice from './routes/practice.tsx';
-import AccountContainer from "containers/account/AccountContainer";
-import Report from './routes/report.tsx';
-import PdfReport from './routes/pdfReport.tsx';
+import Practice from './pages/practice.tsx';
+import Report from './pages/report.tsx';
+import PdfReport from './pages/pdfReport.tsx';
+import Account from './pages/account.tsx';
 
 
 const router = createBrowserRouter(
@@ -25,7 +25,7 @@ const router = createBrowserRouter(
         element={<App />} >
         <Route path="/" element={<Home />} />
         <Route path="/practice" element={<Practice />} />
-        <Route path="/account" element={<AccountContainer />} />
+        <Route path="/account" element={<Account />} />
         <Route path="/report/:id" element={<Report />} />
         <Route path="/report/pdf/:id" element={<PdfReport />} />
       </Route>
