+++ src/pages/account.tsx
@@ -1,0 +1,11 @@
+import AccountContainer from "containers/account/AccountContainer";
+
+
+export default function Account() {
+
+  return (
+    <>
+      <AccountContainer />
+    </>
+  )
+}
