+++ src/pages/practice.tsx
@@ -1,9 +1,9 @@
 import PracticeContainer from "containers/practice/PracticeContainer"
 
 // The practice route will render the practice module.
-export default function Practice () {
-  
+export default function Practice() {
+
   return (
-    <PracticeContainer/>
+    <PracticeContainer />
   )
 }
